requirements.txt
Django==2.1
celery==4.2.1
djangorestframework==3.11.0
pika==1.1.0